Initial modifications to fix MacOS install bug
This has a few changes to enable more testing by the bioconda CI system
as well as fix the MacOS bug.

The use of an environment variable (CONDA_ROOT) is determined by the
inclusion of the DEFAULT_DB_DIR preprocessor macro. That should be
changed to a flag such as CONDA_BUILD.